+++ e2e/e2etests/test_bitcoin_std_deposit_and_call.go
@@ -7,45 +7,47 @@ import (
 
 	"github.com/zeta-chain/node/e2e/runner"
 	"github.com/zeta-chain/node/e2e/utils"
+	"github.com/zeta-chain/node/pkg/memo"
 	testcontract "github.com/zeta-chain/node/testutil/contracts"
 	crosschaintypes "github.com/zeta-chain/node/x/crosschain/types"
 	zetabitcoin "github.com/zeta-chain/node/zetaclient/chains/bitcoin"
 )
 
-func TestBitcoinDepositAndCall(r *runner.E2ERunner, args []string) {
-	// ARRANGE
-	// Given BTC address
+func TestBitcoinStdMemoDepositAndCall(r *runner.E2ERunner, args []string) {
+	// setup deployer BTC address
 	r.SetBtcAddress(r.Name, false)
 
-	// Given "Live" BTC network
+	// start mining blocks if local bitcoin
 	stop := r.MineBlocksIfLocalBitcoin()
 	defer stop()
 
-	// Given amount to send
+	// parse amount to deposit
 	require.Len(r, args, 1)
 	amount := parseFloat(r, args[0])
-	amountTotal := amount + zetabitcoin.DefaultDepositorFee
-
-	// Given a list of UTXOs
-	utxos, err := r.ListDeployerUTXOs()
-	require.NoError(r, err)
-	require.NotEmpty(r, utxos)
 
 	// deploy an example contract in ZEVM
 	contractAddr, _, contract, err := testcontract.DeployExample(r.ZEVMAuth, r.ZEVMClient)
 	require.NoError(r, err)
-	r.Logger.Info("Bitcoin: Example contract deployed at: %s", contractAddr.String())
 
-	// ACT
-	// Send BTC to TSS address with a dummy memo
-	data := []byte("hello satoshi")
-	memo := append(contractAddr.Bytes(), data...)
-	txHash, err := r.SendToTSSFromDeployerWithMemo(amountTotal, utxos, memo)
-	require.NoError(r, err)
+	// create standard memo with [receiver, payload]
+	memo := &memo.InboundMemo{
+		Header: memo.Header{
+			Version:     0,
+			EncodingFmt: memo.EncodingFmtCompactShort,
+			OpCode:      memo.OpCodeDepositAndCall,
+		},
+		FieldsV0: memo.FieldsV0{
+			Receiver: contractAddr,
+			Payload:  []byte("hello satoshi"),
+		},
+	}
+
+	// deposit BTC with standard memo
+	txHash := r.DepositBTCWithAmount(amount, memo)
 
 	// wait for the cctx to be mined
 	cctx := utils.WaitCctxMinedByInboundHash(r.Ctx, txHash.String(), r.CctxClient, r.Logger, r.CctxTimeout)
-	r.Logger.CCTX(*cctx, "bitcoin_deposit_and_call")
+	r.Logger.CCTX(*cctx, "bitcoin_std_memo_deposit_and_call")
 	utils.RequireCCTXStatus(r, cctx, crosschaintypes.CctxStatus_OutboundMined)
 
 	// check if example contract has been called, 'bar' value should be set to amount