+++ x/crosschain/types/message_vote_inbound.go
@@ -22,6 +22,13 @@ const MaxMessageLength = 10240
 // InboundVoteOption is a function that sets some option on the inbound vote message
 type InboundVoteOption func(*MsgVoteInbound)
 
+// WithMemoRevertOptions sets the revert options for inbound vote message
+func WithRevertOptions(revertOptions RevertOptions) InboundVoteOption {
+	return func(msg *MsgVoteInbound) {
+		msg.RevertOptions = revertOptions
+	}
+}
+
 // WithZEVMRevertOptions sets the revert options for the inbound vote message (ZEVM format)
 // the function convert the type from abigen to type defined in proto
 func WithZEVMRevertOptions(revertOptions gatewayzevm.RevertOptions) InboundVoteOption {