+++ changelog.md
@@ -21,6 +21,7 @@
 * [2987](https://github.com/zeta-chain/node/pull/2987) - add non-EVM standard inbound memo package
 * [2979](https://github.com/zeta-chain/node/pull/2979) - add fungible keeper ability to lock/unlock ZRC20 tokens
 * [3012](https://github.com/zeta-chain/node/pull/3012) - integrate authenticated calls erc20 smart contract functionality into protocol
+* [3025](https://github.com/zeta-chain/node/pull/3025) - standard memo for Bitcoin inbound
 
 ### Refactor